Use reflection to invoke new MiniAccumuloCluster classes when present
>=Accumulo-1.6.1 was updated to use thrift-0.9.1. Hive-exec bundles
thrift-0.9.0 inside. This is very problematic. Thift has, and
likely never will, support this gracefully.

The new MiniAccumuloCluster classes also has some new methods
which allow for control of the classpath used to run the minicluster.
We can remove the hive-exec jar from the MAC classpath which
alleviates the problem of multiple thrift jars on the classpath.